requirements.txt
Django>=2.1.3,<2.2.0
djangorestframework>=3.9.1,<3.10.0
flake8>=3.6.0,<3.7.0
markdown2>=2.3.7,<2.4.0
psycopg2-binary>=2.7.7,<2.8.0
